Ya quedo, al parecel ppid es extrem y no ./init

+++ init.c
@@ -3,13 +3,9 @@
 #include <unistd.h>
 #include <signal.h>
 #include <sys/wait.h>
-pid_t procesos[6];
-int finish=0;
-void sigquit();
 
 int main(){
-    signal(SIGQUIT, sigquit);
-    pid_t ended;
+    pid_t procesos[6],ended;
     int i=0;
     for (int i = 0; i < 6; i++)
     {
@@ -21,37 +17,36 @@ int main(){
     }
     //codigo para revisar que siempre haya 6 getty
     int status;
-    while(!finish){ 
+    while(1){ 
         ended = waitpid(-1, &status, 0);
         printf("Ended: %d, Status: %d\n",ended,WEXITSTATUS(status));
         for (int i = 0; i < 6; i++) // Constantemente, revisar todos los hijos
         {
             
             if(ended == procesos[i]){ // Si un hijo finalizo... (se cerro la ventana)
-            if (!finish)
-            {
-                procesos[i]=fork(); // Crear uno nuevo y guardar su pid
+                                      // Verificar si se cerro por un shutdown, en caso de que si, cerrar todos los procesos
+                
+                if(WEXITSTATUS(status)==3){
+                    for (int i = 0; i < 6; i++)
+                    {
+                        if(procesos[i]!=ended){
+                            kill(procesos[i], SIGQUIT);
+                        }
+                    }
+                    while(wait(NULL)>0);
+                    return(0);
+                }else{
+                    procesos[i]=fork(); // Crear uno nuevo y guardar su pid
                 if(procesos[i]==0){
                     //execlp("sh","sh","xterm -e ./getty",NULL);
                     execlp("xterm", "xterm", "-e", "./getty", NULL); // Reemplazarlo por la ventana de xterm (xterm -e ./getty)
                     exit(0); //dado caso que no se ejecute para que no continue
                 }
-            }
-                                      // Verificar si se cerro por un shutdown, en caso de que si, cerrar todos los proceso
-
+                }
+                
             }
         }
         
     }
     
-}
-
-
-void sigquit() {
-    for (int i = 0; i < 6; i++) {
-        printf("SIGNAL RESIVED");
-        sleep(5);
-        kill(procesos[i], SIGQUIT);
-    }
-    exit(0);
 }
